[Refactor] Change filename and add a new function
